+++ LiquidDiffract/gui/main_widget.py
@@ -212,7 +212,7 @@ class MainContainer(QWidget):
         if (self.optim_ui.optim_config_widget.optim_options_gb.opt_check.isChecked()) & (self.optim_ui.optim_config_widget.optim_options_gb.isChecked()):
             self.results_ui.data['rho'] = self.optim_ui.data['refined_rho']
         else:
-            self.results_ui.data['rho'] = np.float(self.optim_ui.optim_config_widget.composition_gb.density_input.text())
+            self.results_ui.data['rho'] = np.float64(self.optim_ui.optim_config_widget.composition_gb.density_input.text())
         if self.optim_ui.optim_config_widget.optim_options_gb.isChecked():
             self.results_ui.data['int_func'] = self.optim_ui.data['impr_int_func']
             self.results_ui.data['sq_x'] = self.optim_ui.data['impr_iq_x']
